% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_rtf.R
\name{arrange_rtf}
\alias{arrange_rtf}
\title{Arranger RTF files using a Shiny UI.}
\usage{
arrange_rtf(path_save, copy_local = TRUE)
}
\arguments{
\item{path_save}{Optional. Desired location of the lookup file. Users
may specify a save path where \code{order.csv} will be created, with columns
\verb{File Name}, \verb{Table Name}, and \code{FilePath}. \code{FilePath} will include a dynamically
created temporary directory. A dataframe \code{rtf_order} will be created in the
workspace with columns \verb{File Name}, \verb{Table Name}, and \code{FilePath} regardless of
the \code{path_save} value.}
}
\value{
A Shiny app that can be run within RStudio or in a browser. The function
does not return an object, it instead creates a dataframe in
the workspace, and, optionally, a csv file at \code{path_save}, both of which
will have the same data.
}
\description{
\code{arrange_rtf} is used to create a lookup file that places RTF files
in a desired order. The user "uploads" the RTF files, which are copied to
a local temporary directory. The user can then drag/drop rows to order
the RTF files as desired.
}
\details{
Upon execution of \code{arrange_rtf}, users should click the Selection
button and select the RTF files to be combined. A text appears describing
what dataframe was created, and potentially where a csv file was saved.
Additionally, a table appears with RTF file details. This table can be manipulated
by clicking on the row number on the left and dragging the row to change the order.
This is especially helpful when the table names differ from the file name, or
a specific order is desired.

The \code{FilePath} column shows where the files now reside, inside a temporary folder.
This temporary folder may not persist once RStudio is closed, so it is advised
to use the row order information within the current R session.

It is possible to click the \code{Selection} button multiple times, which is
especially useful if files in multiple locations need to be combined.
}
\examples{
\dontrun{
# Run with a path input
arrange_rtf("/my/save/path")

# Run without a path input - same result:
arrange_rtf()
}

}
