% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_rtf.R
\name{parse_caption_rtf}
\alias{parse_caption_rtf}
\title{RTF caption parser}
\usage{
parse_caption_rtf(x)
}
\arguments{
\item{x}{character. Should be a multi-vector result of using \code{striprtf::read_rtf}
to load an RTF file.}
}
\value{
A character string.
}
\description{
\code{parse_caption_rtf} is used to parse out the caption from an RTF file
created by \code{r2rtf}.
}
\details{
\code{parse_caption_rtf} tries to parse out the table or figure caption from
an RTF. The function finds the first instance of figure or table in the RTF.
Next, it selects the lines containing figure or table and minimal clean-up is
attempted to present a well-formatted result.
}
\examples{
\dontrun{
# Create an RTF
file <- tempfile(fileext = ".rtf")

file1 <- head(iris) \%>\%
  rtf_title(title = "Table 1. My table") \%>\%
  rtf_body() \%>\%
  rtf_encode() \%>\%
  write_rtf(file)

x <- striprtf::read_rtf(file)

parse_caption_rtf(x)
}

}
